Add Home component
This commit adds a new component which is meant to be displayed at the root before the user logs in. As soon as a user is present in the state, the component redirects to `/site`.

The component has a bunch of static HTML, much of which does not elegantly translate to JSX, so the HTML is stored alongside the component in `home.html`. The HTML is loaded as a string using stringify and then set when the component is rendered.