Fix intermittent GitHub auth test failure
There was an intermittent GitHub auth test failure that I was able to track down to an issue with the nock for the GitHub access token request. It appears that the `passport-oauth` strategy, which the `passport-github` strategy is built on occassionally sends an additional `redirect-uri` param. This appears to depend on the URL that `supertest` assigns the app when it sends the request. When this happens, the request body for the nock does not match.

This commit changes the request body matching the access token nock to use a function instead of an object so an exact match is not required and the additional params will no longer break the test.