Save session before redirecting on authentication
This commit addresses a race condition in the session store.

We change the session when we authenticate, then redirect before saving. This meant that the next request could load the old session. The old session would not be marked as authenticated, so auth would fail. I haven't observed this bug on staging or prod, but have seen it occasionally running locally.

This commit fixes the issue by calling `session.save` and waiting for the session to save before triggering the redirect.