Move published files list out of published branch API into published files API
This commit adds a new route which is used to list the files published to a branch. This new route corresponds to a "published files" API and is:

```
/v0/site/{site_id}/published-branch/{branch}/published-file
```

This renders a list of objects that correspond to the files published to S3 for the given branch.

This commit also removes this information from the published branch `findOne` action.