Integrate published files API with the frontend
This commit uses the new published files API to display published files in the frontend. It adds a new sets of actions and a new reducer and moves the existing functionality out of the published branches reducer/actions. Additionally, it renames the `SitePublishedBranch` component to `SitePublishedFilesTable` and changes it to work with the new reducer / actions.