Display loading indicator when waiting for user data to load
This commit fixes a number of bugs that occured around the user loading states.

When the user data was loading, the site displayed a white screen instead of an indicator. This commit adds the indicator to the `<App/>` component while the user fetch is running.

When the user fetch was triggered, the `USER_FETCH_STARTED` action was not dispatched because the function was not actually invoked. This commit fixes that by adding the parens to invoke the function.

This commit removes the loading indicator from the Home component, since displaying it will be handled by the aApp component.

Finally, this commit adds the loading indicator to some of the components that were previously displaying loading messages.