Next the build find action under sites
This commit changes the build's find action to be available at `/v0/site/:site_id/build` instead of `/v0/site`. It now serves a list of builds for the given site, instead of all of the user's builds.

This commit also removes the `builds` and `users` arrays from the site response objects. This will make the response faster since it doesn't need to query and serialize this data.