Add publishedAt attribute to site
Removing the `builds` array from the site response object means that the frontend can no longer iterate over the list of builds associated with the site to find the most recent published build. In order to prevent the frontend from sending a request to query this data for each site, this commit adds a `publishedAt` column to site. This is updated when `build.completeJob` is called without an error.

This attribute is present in the API response where the frontend can reach it to display in the site list.