Cleanup up site state
This commit cleans up the site state in preparation for adding a loading state to the site list. This commit removes unused action and reducer code and pulls some code out of the components. A lot of the code removed in this commit is left behind from removing the editor.