Add published branches loading state to the frontend
This commit changes the published branches redux state so it has an `isLoading` prop and `data` prop. This allows us to know if published branches are being loaded. The changes to the state are used to add a loading indicator to the published branch list if published branches are being downloaded.