Fix order of build array
The build array that was rendered by the build API was showing builds in an incorrect order. This was confusing because the builds query set the order in the correct manner.

It turned out the issue was that the build serializer was running another query that did not specify the order. This commit moves the code to set the order into the query in the build serializer.

Ref #898