Add signedInAt attribute to user model
This commit adds a signedInAt attribute to the user model. This attribute is touched every time a user signs in. This is helpful because it can be used to fetch the most recently signed in user for a site, who is most likely to have a valid GitHub access token.