Fix issue where Federalist was creating org sites instead of user sites
When creating a repo, the API request is different for creating the repo for a user versus for an org. Because of this reason, we need to compare the repo owner string to the username to determine what kind of request to send.

Federalist downcases the repo name for a new site when creating a site from a template. Because of this, if the user has capital letters in their username, the comparison will fail where it should succeed.

For example, if a user named `FederalistUser` tries to create a template site, the owner name will be `federalistuser`. When the owner string is compared to the user's username, the comparison will fail and a request will be sent to the new org repo API endpoint, resulting in a 404 since the `federalistuser` org does not exist.

This commit fixes the issue by downcasing the username and owner before the comparison. Additionally, a regression test is added for this case.

Ref #957