Add site wide error message
This commit adds the ability to display a site wide error message. This can be done via the `federalist-site-wide-error` user provided service which takes the following shape:

```json
{
  "DISPLAY": true,
  "HEADING": "There has been a site wide error",
  "BODY": "This error is breaking things. <a href='www.example.gov'>Click here</a> for more info."
}
```

The error is displayed at the top of the site. The body is displayed as unescaped HTML to allow for the inclusion of links, bulleted lists, etc

Ref #900