Fix issue where SiteRemover caused an internal server error
This commit fixes 2 issues with the S3SiteRemover service, and adds regression test to prevent the issues in the past.

Both issues relate to issues with how the `delete-objects` API request was sent to the S3 API.

The first issue was that an empty request (i.e. a request to delete 0 objects) would cause a the S3 API to respond with a 500 which would propagate to the S3SiteRemover. This commit adds a guard clause to the top of the `deleteObjects` function to prevent this from happening.

The second issue happened because the `delete-objects` API requests does not allow for the delete of more than 1000 objects at a time. This commit changes `deleteObjects` function so that it returns a Promise that recursively deletes objects 1000 at a time. This means at most 1000 objects are deleted at a time and no deletion requests are happening simultaneously threatening to break AWS's rate limit.

Ref #990