change p font-size to 15px
The font face we use, Merriweather, looks weird in fractional sizes
on Windows, as reported in #1124.

This PR changes the font size to a whole number value of 15px, which
is nearly un-noticeably (to me, anyway) smaller than the computed 15.3px
size we previously had when it was set to 0.9em.