improve Add Site UX
First show the owner and repository fields in the Add Site flow.
If a site exists in Federalist (that the user can access), then add the current
user to that site.
Otherwise, present the user with the engine and primary branch fields so
that a completely new site will be added to Federalist.

also:
- extract AddRepoSiteForm and use redux-form for it
- create new back end method to add a user to a Federalist site based on owner/repo
- lots of linting and tests