Lowercase org name before organization check
The org name that we get from the GitHub API can be uppercase or
lowercase, but we lowercase the "owner" param when we consume it in the
SiteCreator. In order to properly compare the org name with the repo
owner name, we need to down case the org name we receive from GitHub.